middelwares & CRUD users modified to admin

+++ controllers/groups.controller.js
@@ -1,4 +1,5 @@
 const Group = require ('../models/group.model');
+const mongoose = require ('mongoose');
 
 module.exports.create = (req,res,next)=> {
     const group = new Group (req.body);
@@ -10,14 +11,15 @@ module.exports.create = (req,res,next)=> {
 }
 
 module.exports.list = (req,res,next) => {
-    Group.find()
+    Group.find({admin:req.user.id})
     .then(groups => res.json(groups))
     .catch(error => next(error));
 
 }
 
 module.exports.select = (req,res,next) => {
-    Group.findById(req.params.id)
+    Group.findById({admin:req.params.userId, _id : req.params.id}
+)
     .then(group => {
         if(!group){
             throw createError(404,'group not found')
@@ -28,8 +30,8 @@ module.exports.select = (req,res,next) => {
 }
 
 module.exports.update = (req,res,next) => {
-    Group.findById(req.params.id)
-    .then(gruop =>{
+    Group.findById({admin:req.params.userId,_id: req.params.id})
+    .then(group =>{
         if (!group){
             throw createError(404, 'group not found')
         }else{
@@ -42,9 +44,10 @@ module.exports.update = (req,res,next) => {
 }
 
 module.exports.delete = (req,res,next) => {
-    Group.findByIdAndRemove(req.params.id) 
+    Group.findByIdAndRemove({admin:req.params.userId ,_id: req.params.id}) 
     .then(() => {
         res.status(204).json()
     })
     .catch(error => next (error));
 }   
+
